requirements.txt
discord.py
python-decouple
requests-html
bs4
selenium
requests 
time